p146 LRU

+++ scripts/add_problem.sh
@@ -10,6 +10,17 @@ problem_number=$1
 src_file="$SCRIPT_DIR/../src/p_$(printf "%04d" $problem_number).ts"
 test_file="$SCRIPT_DIR/../tests/p_$(printf "%04d" $problem_number).test.ts"
 
+# Check if files already exist
+# if exists, ask if user wants to overwrite
+if [ -f $src_file ] || [ -f $test_file ]; then
+    read -p "Files already exist. Do you want to overwrite? (y/n) " -n 1 -r
+    echo
+    if [[ ! $REPLY =~ ^[Yy]$ ]]; then
+        echo "Aborting..."
+        exit 1
+    fi
+fi
+
 # Create src file with basic content
 cat << EOF > $src_file
 // LeetCode - $problem_number
